small updates
I'm done playing FFXI for a while, since I was getting bored.  I finished a mythic (vajra) and three relics (excal, apoc, yoichi) while I was back playing... so that's pretty cool.  But that's about the extent of what I cared to do solo, and I had no desire to rejoin a group.

So I'm done with FFXI for a while, once again.